+++ packages/wallets/test/unit/wallet-util.ts
@@ -107,9 +107,9 @@ describe("WalletUtil", () => {
 
         const jsonBody = { "method": "getVisibility", "timestamp": 1582196988554 }
 
-        const signature = await JsonSignature.sign(jsonBody, wallet)
+        const signature = await JsonSignature.signMessage(jsonBody, wallet)
 
-        expect(JsonSignature.isValid(signature, computePublicKey(wallet.publicKey, true), jsonBody)).to.be.true
-        expect(JsonSignature.isValid(signature, wallet.publicKey, jsonBody)).to.be.true
+        expect(JsonSignature.isValidMessage(jsonBody, signature, computePublicKey(wallet.publicKey, true))).to.be.true
+        expect(JsonSignature.isValidMessage(jsonBody, signature, wallet.publicKey)).to.be.true
     })
 })