+++ packages/data-models/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @vocdoni/data-models - Changelog
 
+## 1.15.3
+
+- Minor typedef change
+
+## 1.15.2
+
+- Adding `wrapRawTransaction`
+
 ## 1.15.0
 
 - Adding support for anonymous voting (zkSnarks)